refactor: Update TracesList component to use FieldType.DateTime for "Seen At" field
This commit updates the TracesList component to use the FieldType.DateTime instead of FieldType.Date for the "Seen At" field. This change ensures that the field accurately represents the date and time of when the trace was seen, improving the accuracy and clarity of the telemetry data.
